donut/requirements.txt
aiohttp
aiosignal
anyio
async-timeout
attrs
bcrypt
certifi
cffi
charset-normalizer
click
config
contourpy
cryptography
cycler
datasets
dill
fastapi
ffmpy
filelock
fonttools
frozenlist
fsspec
gradio
h11
httpcore
httpx
huggingface-hub
idna
Jinja2
joblib
kiwisolver
lightning-utilities
linkify-it-py
markdown-it-py
MarkupSafe
matplotlib
mdit-py-plugins
mdurl
multidict
multiprocess
munch
nltk
numpy
nvidia-cublas-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
orjson
packaging
pandas
paramiko
Pillow
protobuf
pyarrow
pycparser
pycryptodome
pydantic
pydub
PyNaCl
pyparsing
python-dateutil
python-multipart
pytorch-lightning
pytz
PyYAML
regex
requests
responses
rfc3986
ruamel.yaml
ruamel.yaml.clib
sconf
six
sniffio
starlette
tensorboardX
timm
tokenizers
torch
torchmetrics
torchvision
tqdm
transformers
typing-extensions
uc-micro-py
urllib3
uvicorn
websockets
xxhash
yarl
zss
sentencepiece
python-dotenv
minio
omegaconf